+++ app/src/main/java/com/ping/android/activity/GroupProfileActivity.java
@@ -209,7 +209,7 @@ public class GroupProfileActivity extends CoreActivity implements View.OnClickLi
     private void openPicker() {
         String profileFileFolder = getExternalFilesDir(null).getAbsolutePath() + File.separator +
                 "profile" + File.separator + currentUser.key;
-        long timestamp = System.currentTimeMillis() / 1000L;
+        double timestamp = System.currentTimeMillis() / 1000L;
         String profileFileName = "" + timestamp + "-" + currentUser.key + ".png";
         String profileFilePath = profileFileFolder + File.separator + profileFileName;
         imagePickerHelper = ImagePickerHelper.from(this)
